# packages/backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/create-user": {
      "post": {
        "operationId": "AuthController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. Only admins can create users."
          },
          "409": {
            "description": "Conflict. User with this email already exists."
          }
        },
        "summary": "Create a new user (Admin Only)",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful, returns tokens and user info."
          },
          "401": {
            "description": "Invalid credentials."
          }
        },
        "summary": "Log in a user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns current user profile.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refreshToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully refreshed access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshTokenResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Access Denied (invalid or missing refresh token)."
          }
        },
        "summary": "Refresh access token using refresh token",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully logged out."
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Log out current user (invalidates refresh token)",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/password/reset-request": {
      "post": {
        "operationId": "AuthController_requestPasswordReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/password/reset": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
...
  },
  "info": {
    "title": "Supplement Management API",
    "description": "API documentation for the Supplement Management System",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Full name of the user",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "description": "Email address of the user",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "description": "User password, at least 8 characters long",
            "example": "password123"
          },
          "role": {
            "type": "string",
            "description": "Role of the user",
            "enum": [
              "admin",
              "consultant"
            ],
            "example": "consultant"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "role"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Unique user identifier",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "User's full name",
            "example": "Admin User"
          },
          "email": {
            "type": "string",
            "description": "User's unique email address",
            "example": "admin@example.com"
          },
          "role": {
            "type": "string",
            "description": "User's role",
            "enum": [
              "admin",
              "consultant"
            ],
            "example": "admin"
          },
          "status": {
            "type": "string",
            "description": "User's account status",
            "enum": [
              "active",
              "inactive"
            ],
            "example": "active"
          },
          "lastActive": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of the user's last activity",
            "example": "2025-05-07T13:05:44.879Z",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when the user was created",
            "example": "2025-05-06T16:20:51.170Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of the last update to the user profile",
            "example": "2025-05-07T11:05:44.931Z"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "role",
          "status",
          "lastActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email for login",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "description": "User password for login",
            "example": "password123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token used to obtain a new access token",
            "example": "anotherlongrandomstringrefreshtoken67890"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "RefreshTokenResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "New access token generated after successful refresh",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFkbWluQGV4YW1wbGUuY29tIiwic3ViIjoxLCJyb2xlIjoiYWRtaW4iLCJpYXQiOjE3NDY2MjM1MDAsImV4cCI6MTc0NjYyNzEwMH0.abcdef1234567890"
          }
        },
        "required": [
          "accessToken"
        ]
      },
      "PasswordResetRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email address to send password reset link to",
            "example": "user@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "PasswordResetDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Password reset token received via email",
            "example": "longrandomstringtoken12345"
          },
          "password": {
            "type": "string",
            "description": "New password, at least 8 characters long",
            "example": "newSecurePassword123"
          },
          "passwordConfirmation": {
            "type": "string",
            "description": "Confirmation of the new password, must match password field",
            "example": "newSecurePassword123"
          }
        },
        "required": [
          "token",
          "password",
          "passwordConfirmation"
        ]
      },
...
    }
  }
}